### 2018-01-31 54bce08
TIKA-2556: Swap out com.tdunning:json for com.github.openjson:openjson to avoid jar conflicts.
