### 2018-01-31 27a9c40
TIKA-2547 -- RFC822 with multipart/mixed, first text element should be treated      as the main body of the email, not an attachment.
