### 2018-01-31 6829643
TIKA-2547: RFC822 with multipart/mixed, first text element should be treated as the main body of the email, not an attachment.
