### 2018-02-05 5244cde
TIKA-1983 -- remove deprecated server from tika-app in Tika 2.0
