+++ tika-app/src/main/java/org/apache/tika/cli/TikaCLI.java
@@ -37,8 +37,6 @@ import java.io.PrintWriter;
 import java.io.UnsupportedEncodingException;
 import java.io.Writer;
 import java.lang.reflect.Field;
-import java.net.ServerSocket;
-import java.net.Socket;
 import java.net.URI;
 import java.net.URL;
 import java.nio.charset.Charset;
@@ -350,8 +348,6 @@ public class TikaCLI {
 
     private boolean pipeMode = true;
 
-    private boolean serverMode = false;
-
     private boolean fork = false;
 
     private boolean prettyPrint;
@@ -453,8 +449,9 @@ public class TikaCLI {
             prettyPrint = true;
         } else if (arg.equals("-p") || arg.equals("--port")
                 || arg.equals("-s") || arg.equals("--server")) {
-            serverMode = true;
-            pipeMode = false;
+            throw new IllegalArgumentException(
+                    "As of Tika 2.0, the server option is no longer supported in tika-app.\n"+
+                    "See https://wiki.apache.org/tika/TikaJAXRS for usage.");
         } else if (arg.startsWith("-c")) {
             networkURI = new URI(arg.substring("-c".length()));
         } else if (arg.startsWith("--client=")) {
@@ -463,9 +460,7 @@ public class TikaCLI {
             pipeMode = false;
             configure();
             
-            if (serverMode) {
-                new TikaServer(Integer.parseInt(arg)).start();
-            } else if (arg.equals("-")) {
+            if (arg.equals("-")) {
                 try (InputStream stream = TikaInputStream.get(
                         new CloseShieldInputStream(System.in))) {
                     type.process(stream, System.out, new Metadata());
@@ -533,7 +528,7 @@ public class TikaCLI {
     }
     private void usage() {
         PrintStream out = System.out;
-        out.println("usage: java -jar tika-app.jar [option...] [file|port...]");
+        out.println("usage: java -jar tika-app.jar [option...] [file...]");
         out.println();
         out.println("Options:");
         out.println("    -?  or --help          Print this usage message");
@@ -541,7 +536,6 @@ public class TikaCLI {
         out.println("    -V  or --version       Print the Apache Tika version number");
         out.println();
         out.println("    -g  or --gui           Start the Apache Tika GUI");
-        out.println("    -s  or --server        Start the Apache Tika server");
         out.println("    -f  or --fork          Use Fork Mode for out-of-process extraction");
         out.println();
         out.println("    --config=<tika-config.xml>");
@@ -607,12 +601,6 @@ public class TikaCLI {
         out.println("    Apache Tika GUI. You can drag and drop files from");
         out.println("    a normal file explorer to the GUI window to extract");
         out.println("    text content and metadata from the files.");
-        out.println();
-        out.println("- Server mode");
-        out.println();
-        out.println("    Use the \"--server\" (or \"-s\") option to start the");
-        out.println("    Apache Tika server. The server will listen to the");
-        out.println("    ports you specify as one or more arguments.");
         out.println();
         out.println("- Batch mode");
         out.println();
@@ -1114,66 +1102,6 @@ public class TikaCLI {
         }
     }
 
-    private class TikaServer extends Thread {
-
-        private final ServerSocket server;
-        private final int port;
-        public TikaServer(int port) throws IOException {
-            super("Tika server at port " + port);
-            server = new ServerSocket(port);
-            this.port = port;
-        }
-
-        @Override
-        public void run() {
-            System.out.println("Successfully started tika-app's server on port: "+port);
-            System.err.println("WARNING: The server option in tika-app is deprecated and will be removed ");
-            System.err.println("by Tika 2.0 if not shortly after Tika 1.14.");
-            System.err.println("Please migrate to the JAX-RS tika-server package.");
[…]
