### 2018-02-05 7d22ba2
TIKA-2564 -- wrap embedded stream in a stream that supports mark/reset in --extract option in tika-app
