### 2018-02-05 504ba00
TIKA-2564 -- wrap embedded stream in a stream that supports mark/reset in --extract option in tika-app
