### 2018-02-07 0e5fded
TIKA-2559: Extract language metadata item from PDF files via Matt Sheppard.
