### 2018-02-07 4645e43
TIKA-2559: Extract language metadata item from PDF files via Matt Sheppard.
