+++ CHANGES.txt
@@ -5,6 +5,8 @@ Release 2.0.0 - ???
 
    Other changes
 
+   * Extract language metadata item from PDF files via Matt Sheppard (TIKA-2559)
+
    * RFC822 with multipart/mixed, first text element should be treated
      as the main body of the email, not an attachment (TIKA-2547).
 
