### 2018-02-07 5314bc4
update changes for TIKA-2559
