### 2018-02-09 d9be32c
TIKA-2569 -- Extract text from grouped text boxes in PPT.
