### 2018-02-09 4c510d6
TIKA-2569 -- Extract text from grouped text boxes in PPT.
