### 2018-02-09 cef8494
TIKA-2571 -- re-throw SecurityException in ZipContainerDetector
