### 2018-02-09 #191 @noamross
### Editor checks:

---

#### Editor comments

Thanks for your submission @predict-r! And sorry for the slow reply, this fell through the cracks for a week.  This is a good fit and I'll go ahead and find reviewers.

Here are the results of our automated tests via `goodpractice::gp()`. You have good code coverage (100% not required, but this helps rviewers see where they might look for edge cases that need coverage.)  Most of the other stuff should be straightforward and I'll go ahead and look for reviewers.  You can make those changes either before or after you get your first-round reviews.

I suspect you'll have trouble getting a >50MB `.jar` file onto CRAN. I'll leave this until we have a review from Java-experienced reviewer, but I suspect you may have to pursue an approach such as packaging the `.jar` file in a separate package and listing it as a dependency ([see CRAN polices](https://cran.r-project.org/web/packages/policies.html)),  file or having a function like `install_tika()` in the package that downloads the file before the package is functional.

```
── GP rtika ─────────────────────────────────────

It is good practice to

    R/tika_html.R:12:NA
    R/tika_json.R:17:NA
    R/tika_text.R:13:NA
    R/tika_xml.R:13:NA
    R/tika.R:128:NA
    ... and 7 more lines

  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users and developers are used it and it is easier to read your code for them if you use '<-'.

    inst/doc/text_processing.R:121:11

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines shorter than 80 characters

    inst/doc/text_processing.R:12:1
    inst/doc/text_processing.R:58:1
    inst/doc/text_processing.R:62:1
    inst/doc/text_processing.R:84:1
    inst/doc/text_processing.R:106:1
    ... and 56 more lines

  ✖ omit trailing semicolons from code lines. They are not needed and most R coding standards forbid them

    inst/doc/text_processing.R:36:66

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on the input data. Consider using vapply() instead.

    R/tika.R:185:18

  ✖ not use exportPattern in NAMESPACE. It can lead to exporting functions unintendedly. Instead, export functions that constitute the external API of your package.
  ✖ fix this R CMD check NOTE: installed size is 53.3Mb sub-directories of 1Mb or more: java 53.1Mb
─────────────────────────────────────────────
```
---

Reviewers:
Due date:
