### 2018-02-11 #191 @predict-r
Thanks! I'll go ahead and add the badge then. Unless you tell me not to, I'll patch some of the goodpractice issues listed above.

Thanks for pointing out the CRAN situation.  I can create another repository with the Jar, and call it 'tikajar' . That package won't update often, so maybe CRAN will accept it. If not, the package can be left on github and downloaded with devtools.
