### 2018-02-12 fd7ec73
Remove java 8 String.join
