### 2018-02-12 #191 @davidgohel
Hi,

This is a first set of comments, I have only reviewed the DESCRIPTION file for now. If something is not clear, do not hesitate to ask more information.

In the DESCRIPTION file:

* `SystemRequirements` should be shorten to:

    SystemRequirements: java (>= 7)

* I would add a `.onLoad` to check the correct configuration of `rJava` (see
https://cran.r-project.org/doc/manuals/r-release/R-exts.html *Make sure that any version requirement for Java code is both declared in the ‘SystemRequirements’ field and tested at runtime*):

```
#' @import rJava
.onLoad= function(libname, pkgname){

  .jinit()
  jv <- .jcall("java/lang/System", "S", "getProperty", "java.runtime.version")
  reg <- "([0-9]+)\\.([0-9]+)\\.(.*)"
  majorv <- as.integer(gsub(reg, "\\1", jv))
  minorv <- as.integer(gsub(reg, "\\2", jv))
  if(majorv == 1 ){
    if( minorv < 7 ) {
      stop("java version should be at minimum version 7 (>=1.7)")
    }
  }
```

* The field `Depends` says R version should be `>= 3.1.0` but I am not sure it is necessary. Maybe I missed something there; if it is because of `testthat`, that is not necessary.

* The field `URL` is pointing to the github repository. A blogdown site would be a nice thing.

* Fields `Author` and `Maintainer` can be replaced by:

```
Authors@R: c(
    person("Sasha", "Goodman", role = c("aut", "cre"),
    email = "goodmansasha@gmail.com"),
    person("The Apache Software Foundation", role = c("cph"),
    comment = "Apache Tika author") )
```

* I think the package needs a new field `Copyright` and a file `inst/COPYRIGHTS`
(that should contain a copy of tika license).

    Copyright: See file COPYRIGHTS.
