### 2018-02-21 #191 @davidgohel
Hi, here is my review:

- [X] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

*davidgohel*: I tested with 1204 documents (my admin folder). I got 20.354 sec. with 8 threads
and 36.376 with 1 threads. This is really good (it tooks 20 second to get my whole admin
directory scanned and returned as an R object).

I was expecting a faster execution with 8 threads (not 8 times faster but at
least twice minimum). I used the profiling tools of RStudio and could see that the
function spent 1/3 of time executing `.rtika_readFile()` and 1/3 executing `enc2utf8()`.
I think that starting JVM with arg `-Dfile.encoding=UTF-8` instead of using `enc2utf8`
could be a solution (not tested here, but was the solution in past experiences...).

*davidgohel*:

* There is no Code of conduct (but it is only recommanded in the guidelines)
* Authorship: in field `Author` I recommand you the format with `Authors@R: c(person(...)`.
One benefit for example is you will ensure programmatically that the roles are correct.
* Package dependencies: in `R/tika.R` (l179), you are using `utils` but it is not in `Suggests`.
Package `stats` is used in `R/tika.R` (l130) and should be declared in `Imports`.
The way you are using `Suggests` and `requireNamespace` instead of `Imports`
is unconventional to me but I think it is valid.

* `SystemRequirements`: I don't think it will be accepted as is. (2 months ago, R Core Team
standardized java value), it should be `java (>= 7)`. Also CRAN is asking to *double* check the
version of java - something like:

```
.onLoad= function(libname, pkgname){

  .jpackage( pkgname, lib.loc = libname )

  jv <- .jcall("java/lang/System", "S", "getProperty", "java.runtime.version")
  reg <- "([0-9]+)\\.([0-9]+)\\.(.*)"
  majorv <- as.integer(gsub(reg, "\\1", jv))
  minorv <- as.integer(gsub(reg, "\\2", jv))
  if(majorv == 1 ){
    if( minorv < 7 ) {
      stop("java version should be at minimum version 7 (>=1.7)")
    }
  }
}
```

* The run examples are too long to be executed (when *checking the package*).

```
* checking examples ... OK
Examples with CPU or elapsed time > 5s
            user system elapsed
tika      11.477  0.636   8.347
tika_text  5.786  0.328   3.875
tika_json  5.582  0.313   4.163
tika_xml   5.387  0.299   2.787
tika_html  5.224  0.308   2.812
```

As the unit tests are there and cover almost
100%, I suggest you use `\donttest{}` for the 5 examples.

* `R CMD check` did not succeed (via RStudio ckeck button).
I am running R 3.4.3 on OSX High Sierra 10.13.3.
Some tests are failling:

```
OK: 49 SKIPPED: 0 FAILED: 5
1. Error: tika parses multiple local files (@test_tika.R#58)
2. Error: tika outputs NA with a path to nowhere in right order (@test_tika.R#86)
3. Error: tika outputs parsable xml (@test_tika.R#119)
4. Error: tika outputs parsable html (@test_tika.R#135)
5. Error: tika outputs parsable json (@test_tika.R#152)
```

The fun fact is that these are sucessful when launched via RStudio test button... Maybe
that's because of my environment. You can use that bit of code to check the package
with R for OSX on travis:

https://github.com/davidgohel/gdtools/blob/master/.travis.yml#L5

---

Congratulations for that package, it brings someting new
and really useful in the R *data importation* ecosystem.

It works well and is simple to use.
I appreciated the help files. Vignettes are giving
lot of examples and shows how to use the results.

A minor modification might be necessary: starting JVM with option `-Djava.awt.headless=true`,
this should ensure it will run on a server with no graphical environment.

More a wish than a comment... I would have used `rJava` API in order to avoid system calls.
I believe this could cause troubles in corporate environments where IT teams are sometimes
using tools that deny access to command line for some users.
I don't know `tika` but the rJava call should be something like:

```
.jcall("org.apache.tika.wrongbatchclassname","V","main",.jarray(list(...), "java/lang/String")
```

About the size of the jar file, I did not read anything that says it's an
issue (it will be if often updated) but you could ask to CRAN-submissions@R-project.org
or r-package-devel@r-project.org what is the recommanded strategy.

Hope you will find that review useful :)

David
