### 2018-02-21 dd98198
Fix for TIKA-2581 contributed by ewanmellor.
TesseractOCRParserTest.testOCROutputsHOCR fails with Tesseract 4.0.

With 3.x, the output is <span>Happy</span> but with 4.0 the output is
<span><strong>Happy</strong></span>.  Both these seem reasonable to me,
so update the test to accept either of them.
