### 2018-02-21 d1526d0
Fix for TIKA-2582 contributed by ewanmellor.
Tesseract 4.0 includes a change to use form feed characters to separate
pages by default in its text output. Previous versions used no separator
unless you specified the include_page_breaks option.

This confuses any parser that is not expecting the FF.
ODFParserTest.testOO2Metadata fails, because it is expecting the output of
a blank image to be the empty string, but now the FF is there.

I haven't seen any other failures, but I expect that user code will now see
either FF or U+FFFD where they are not expecting it (SafeContentHandler
replaces the F
[…]
