### 2018-02-21 4b96538
Fix for TIKA-2583 contributed by ewanmellor.
Add a section to the readme to mention builds.apache.org.
