### 2018-02-21 9ec8b43
TIKA-2563 -- Extract files embedded in HTML and javascript inside HTML that are stored in the Data URI scheme.
