+++ tika-parsers/src/test/java/org/apache/tika/parser/html/HtmlParserTest.java
@@ -29,6 +29,7 @@ import javax.xml.transform.sax.SAXTransformerFactory;
 import javax.xml.transform.sax.TransformerHandler;
 import javax.xml.transform.stream.StreamResult;
 import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
@@ -39,6 +40,7 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -52,6 +54,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.regex.Pattern;
 
+import org.apache.commons.codec.binary.Base64;
 import org.apache.tika.Tika;
 import org.apache.tika.TikaTest;
 import org.apache.tika.config.ServiceLoader;
@@ -59,6 +62,7 @@ import org.apache.tika.config.TikaConfig;
 import org.apache.tika.detect.AutoDetectReader;
 import org.apache.tika.detect.EncodingDetector;
 import org.apache.tika.exception.TikaException;
+import org.apache.tika.io.IOUtils;
 import org.apache.tika.io.TikaInputStream;
 import org.apache.tika.metadata.Geographic;
 import org.apache.tika.metadata.Metadata;
@@ -1262,6 +1266,36 @@ public class HtmlParserTest extends TikaTest {
                 metadataList.get(1).get(RecursiveParserWrapper.TIKA_CONTENT));
         assertNotContained("cool", metadataList.get(0).get(RecursiveParserWrapper.TIKA_CONTENT));
 
+    }
+
+    @Test
+    public void testDataURI() throws Exception {
+        List<Metadata> metadataList = getRecursiveMetadata("testHTML_embedded_img.html");
+        assertEquals(2, metadataList.size());
+        String content = metadataList.get(0).get(RecursiveParserWrapper.TIKA_CONTENT);
+        assertContains("some content", content);
+        //make sure that you've truncated the data: value
+        assertContains("src=\"data:\"", content);
+        Metadata imgMetadata = metadataList.get(1);
+        assertEquals("image/jpeg", imgMetadata.get(Metadata.CONTENT_TYPE));
+        assertContains("moscow-birds",
+                Arrays.asList(imgMetadata.getValues(TikaCoreProperties.SUBJECT)));
+    }
+
+    @Test
+    public void testDataURIInJS() throws Exception {
+        InputStream is = getClass().getResourceAsStream("/org/apache/tika/parser/html/tika-config.xml");
+        assertNotNull(is);
+        TikaConfig tikaConfig = new TikaConfig(is);
+        Parser p = new AutoDetectParser(tikaConfig);
+        List<Metadata> metadataList = getRecursiveMetadata("testHTML_embedded_img_in_js.html", p);
+        assertEquals(3, metadataList.size());
+        String content = metadataList.get(0).get(RecursiveParserWrapper.TIKA_CONTENT);
+        assertContains("some content", content);
+        Metadata imgMetadata = metadataList.get(1);
+        assertEquals("image/jpeg", imgMetadata.get(Metadata.CONTENT_TYPE));
+        assertContains("moscow-birds",
+                Arrays.asList(imgMetadata.getValues(TikaCoreProperties.SUBJECT)));
     }
 
     @Test
