+++ CHANGES.txt
@@ -1,5 +1,8 @@
 Release 1.18 - ???
 
+   * Extract files embedded in HTML and javascript inside HTML
+     that are stored in the Data URI scheme (TIKA-2563).
+
    * Extract text from grouped text boxes in PPT (TIKA-2569).
 
    * Extract language metadata item from PDF files via Matt Sheppard (TIKA-2559)
