### 2018-02-21 287f5d1
TIKA-2563 -- mods for 1.x branch
