### 2018-02-21 b2ca378
Fix for TIKA-2584 contributed by ewanmellor.
Add TesseractOCRConfig.{add,get}OtherTesseractConfig, plus parsing of
TesseractOCRConfig.properties to extract any key-value pair where the
key has an underscore.

Inside TesseractOCRParser, pass these key-value pairs to Tesseract
using its -c command line option.

This gives a mechanism by which user code can pass arbitrary options
to Tesseract without Tika having to understand them.
