### 2018-02-22 #191 @davidgohel
Awesome, thank you.

There are 2 NOTE(s) left that could be suppressed:

https://travis-ci.org/predict-r/rtika/jobs/344591725#L349

* The first one is because you don't need to keep `rtika.pdf`, R will built it from the `man/` directory.
* the second one is because you are using `packageStartupMessage()` that should not be called from `.onLoad()`, using `warning()` instead should solve that.
