### 2018-02-22 257ee16
TIKA-2580
