### 2018-02-22 35d87d1
TIKA-2580 via Ewan Mellor.
