### 2018-02-22 #191 @predict-r
David, I fixed the issue with the two NOTES.
