### 2018-02-24 #191 @juliasilge
Hello! Here is my review.  :simple_smile:

---

I have had HORRIBLE EXPERIENCES with installation and usage for Java-dependent R packages :scream::scream::scream: so I was full of dread I was asked to review this package. However, I am very happy to report that I was able to install and use this package without any trouble at all. Great job on this package, its test coverage, and usability!

```r
library(rtika)
library(tidyverse)
library(tidytext)

my_path <- "~/Documents/Very Ancient Stuff/Teaching & Astronomy/"

batch <- paste0(my_path, list.files(path = my_path))

text <-
    batch %>%
    tika_text()

tidy_df <-
    data_frame(file = batch,
               text = text) %>%
    unnest_tokens(word, text) %>%
    count(file, word, sort = TRUE) %>%
    anti_join(get_stopwords())
```

```
## Joining, by = "word"
```

```r
tidy_df
```

```
## # A tibble: 3,652 x 3
##    file                                                     word         n
##    <chr>                                                    <chr>    <int>
##  1 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/phy… students   116
##  2 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/phy… writing     68
##  3 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/phy… physics     60
##  4 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/Uni… universe    57
##  5 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/Uni… galaxies    45
##  6 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/Spe… science     43
##  7 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/phy… assignm…    40
##  8 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/phy… class       40
##  9 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/Uni… http        36
## 10 ~/Documents/Very Ancient Stuff/Teaching & Astronomy/Spe… god         33
## # ... with 3,642 more rows
```

```r
metadata <-
    data_frame(raw = batch %>%
                           tika_json()) %>%
    filter(!is.na(raw)) %>%
    mutate(JSON = map(raw, jsonlite::fromJSON, flatten = TRUE)) %>%
    pull(JSON) %>%
    map_df(function(x) { x[1:4] %>% head(1) })

metadata
```

```
##       Application-Name                 Author         Company
## 1 Microsoft PowerPoint           Robert Silge    Robert Silge
## 2      Microsoft Excel           Robert Silge
## 3 Microsoft PowerPoint           Robert Silge UT Southwestern
## 4                 <NA>                   tech            <NA>
## 5 Microsoft PowerPoint Information Technology  Calvin College
## 6 Microsoft PowerPoint               dhaarsma  Calvin College
##   Content-Length    Content-Type        Creation-Date
## 1        5082624            <NA>                 <NA>
## 2          30720            <NA>                 <NA>
## 3        8350720            <NA>                 <NA>
## 4         160175 application/pdf 2001-07-20T19:21:49Z
## 5        1538048            <NA>                 <NA>
## 6        9430528            <NA>                 <NA>
```

#### Issue with R CMD check

When I run `goodpractice:gp()` on the current development version, I get this note about the new **tikajar** dependency:

```
fix this R CMD check NOTE: Namespace in Imports field not imported from: ‘tikajar’
```
I have not built a package that uses a `.jar` file in this way so I do not know how to address this TBH.

#### Code style

I have some suggestions around code style to place this package more squarely within R community norms as expressed, for example, in [Google's R Style Guide](https://google.github.io/styleguide/Rguide.xml) or the [tidyverse style guide](http://style.tidyverse.org/). I don't know how annoying you will find this :flushed: but the value here is that other contributors will be able to read and understand your code more easily. Here are some big picture thoughts:

- Commas at the end of a line, not the beginning; spaces after commas
- [No curly braces around pipes](http://style.tidyverse.org/pipes.html)
- Spaces around equal signs, use `->` for assignment

This is especially important in the `README` and vignette where the code is exposed to more people. I am about to submit a PR with changes to parts of the package but I wasn't able to go thoroughly through all the code today. This is worth it to make this a welcoming R community package.

#### Thoughts on the vignette

The vignette currently focuses on text mining of the text that we can get *out* of this package. Almost the entire vignette focuses on text mining (using base R and some **data.table**) exploratory data analysis, which is not what this package is about; only the very first code chunk uses functions from **rtika** at all. The text mining community in R is active and healthy (I am part of it!) and there are lots of opportunities to learn about how to do text mining once you have text; I would suggest that a vignette for a package like **rtika** is not really the right spot for this kind of analysis. When I am learning how to use a new package, what I want from the vignette is how to use the functions *in that package*. I would love to see, instead, a more detailed exercise of how to use the actual **rtika** functions and what they can do. There is nothing in the vignette about handling metadata from multiple sources, for example. The current material that is in the vignette would be more appropriate for a blog post, or perhaps a secondary, supportive vignette.

I'd also like to address the tone of the vignette a little. I know it is so hard to write for audiences that you aren't quite sure where they are! :smiley: I am 100% in favor of writing accessibly without jargon so that people can get started, but I want to make sure this vignette doesn't come off as condescending. For example, if a user has managed to install the correct version of Java and is attempting to walk through a vignette, they probably know what regular expressions are.
