### 2018-02-25 #191 @noamross
Thanks for your thoughtful review, @juliasilge! I especially concur with your comments on the vignette.

Regarding the `NOTE:` I think the correct field for a package providing `.jar` files is `Depends:`. Perhaps @davidgohel can confirm.
