### 2018-02-25 #191 @predict-r
@juliasilge Thank you for your helpful comments about making the package feel more welcoming. Based on the feedback here, we need a new vignette. It should demonstrate how to use the actual rtika functions and to show what they can do. It is also an opportunity to show the breadth of uses.

I'll merge your pull request, and go through each point in a coming response.

This morning, in the meantime, I'm brainstorming the vignette. If people know a good vignette to study, let me know please!

I'm thinking of a survey of uses and data. It will have more than five examples, each brief, each focused on a distinct case, each using a cool data set, and each following the 'tidy tools manifesto' principles of using pipes and common formats.

The more diverse the examples, the better. I'm thinking :
* handling metadata from multiple sources using tika_json (Julia's example).
* tokenizing and removing stop words with tika_text (similar to Julia's example)
* handling sub-document metadata from the recursive tika_json
* piping XHTML to rvest to get tabular data with tika_html
* maybe an NLP example using tika_text? Something like: https://stanfordnlp.github.io/CoreNLP/
* ? your idea here.

As for data,  my data sources feel too similar to one another and am open to suggestions, especially anything from the Internet Archive (https://archive.org/ ).  My initial ideas:
* a document from the Paradise Papers (  https://ijnet.org/en/blog/paradise-papers-methods-and-tools-investigating-massive-leak )
* a California state legislature committee report on agriculture regulation
* a policy position document from the Internet Archive coming from an association or union.

More diverse data examples feels more welcoming, especially documents appealing to the humanities and arts.

Thank you.
