### 2018-02-25 #191 @juliasilge
Around five example data sets would be plenty for a vignette, I think! Even three small case studies would be good. I would be thoughtful in what tooling you choose to use in the vignettes. For example, if you wanted to use the Stanford CoreNLP in the vignette, then that package would need to be in Suggests, and that package is a BEAR 👿 to get to install on Travis. I know this from experience. I'd try to stay quite lean and simple in what you use in example analysis in the vignettes and keep complex dependencies for demonstration blog posts.

Two examples packages that I think are good to look to for how to handle vignettes in this case are [gutenbergr](https://github.com/ropenscilabs/gutenbergr/tree/master/vignettes) (downloads public domain texts from Project Gutenberg) and [fulltext](https://github.com/ropensci/fulltext/tree/master/vignettes) (searches and downloads full text for journal articles).
