### 2018-02-26 f8b1d9a
TIKA-2580
