### 2018-02-26 520e73f
TIKA-2580
