+++ CHANGES.txt
@@ -5,6 +5,9 @@ Release 2.0.0 - ???
 
    Other changes
 
+   * Extract xlsx files embedded in OLE objects within PPT and PPTX
+     via Brian McColgan (TIKA-2588).
+
    * Extract files embedded in HTML and javascript inside HTML
      that are stored in the Data URI scheme (TIKA-2563).
 
