### 2018-02-26 5e3e910
TIKA-2578 and TIKA-2587 -- Allow for RFC822 detection for files starting with "dkim-" and/or "x-" via Andreas Meier
