### 2018-03-03 #191 @predict-r
@juliasilge, I've implemented changes to address your helpful comments.

> Issue with R CMD check
To  avoid this issue, and make things simpler until CRAN complains, I've put the .jar file back into the 'rtika' package for now. So, the issues with importing 'tikajar' is gone on my end.

 @davidgohel said: "About the size of the jar file, I did not read anything that says it's an
issue (it will be if often updated) "

The size might not be an issue unless the package is updated frequently, and I hope to update it infrequently.

> Code style
The 'styler' package came in handy. I've tried to make the README and vignette use a standard style throughout.

> Thoughts on the vignette
I've written an introductory vignette, and removed the last one. This one focuses on how to use the package functions and process the metadata.

It is on the Github Pages site:
https://predict-r.github.io/rtika/articles/rtika_introduction.html

I tried to keep it light and easy to read, and hope it meets your approval for a package vignette. The site was built with pkgdown::build_site().

By the way, I noticed that the review badge was not loading here, so had to remove it.
