### 2018-03-03 #191 @noamross
Thanks for your response, @predict-r!  @juliasilge, let us know if these changes have addressed your comments and check that box off if so, let us know what else remains if not.  I see @davidgohel already has checked off his box.

FYI, we had a temporary snafu on the service serving badges. It should be back now.,
