### 2018-03-04 cebace9
TIKA-1518: Updated CHANGES file to include description
