### 2018-03-04 2cbca1c
TIKA-1518: Add local docker build based on dockerfile-maven-plugin
