+++ CHANGES.txt
@@ -32,6 +32,9 @@ Release 1.18 - ???
 
    * Upgrade geo-apis to 3.0.1 (TIKA-2535).
 
+   * Added local Docker image build using dockerfile-maven-plugin to allow
+     images to be built from source (TIKA-1518).
+
 Release 1.17 - 12/8/2017
 
   ***NOTE: THIS IS THE LAST VERSION OF TIKA THAT WILL RUN
