### 2018-03-04 d810fba
TIKA-1518: Updated the README and changed image name to tika-server for clarity
