### 2018-03-06 #191 @juliasilge
Ah, I love the new vignette!!! 💕 It is full of so much useful, detailed information. This is going to be a great resource for users. I also am happy to see the other issues in the review addressed; I reviewed the package once more, did R CMD check and `goodpractice::gp()` again, and I have checked off the box in my review.

The only thing significant remaining in the checks is the NOTE about the size of the package.
