### 2018-03-06 #191 @predict-r
I do want this package to be a useful resource, and am so glad you liked the vignette!

> The only thing significant remaining in the checks is the NOTE about the size of the package.
Has anyone seen a package that has an script that downloads data into a package after installation?

I will keep trying, but CRAN never got back to me at all.

Has anybody seen a package that downloads the required resources after package installation? There are Apache mirrors for Tika, such as  http://mirrors.ocf.berkeley.edu/apache/tika/tika-app-1.17.jar .
