### 2018-03-06 #191 @noamross
Thanks all for great reviews and follow-up, @juliasilge @davidgohel and @predict-r.  We can accept the package as-is from our perspective, the size/CRAN issue isn't a precondition on our front, but it makes sense to work out the approach first.  I'll be happy to do a full editors' check if you make major changes on this front.

Courtesy of great friend of rOpenSci Bob Rudis, here's the distribution of `.jar` file sizes on CRAN.  the Tika `.jar` would be larger than any of them, but isn't a different order of magnitude.  So you could go ahead and submit **tikajar** to CRAN and see the what the response is, then submit **rtika** if it goes in.

![uzwxmpjq 1](https://user-images.githubusercontent.com/571752/37013478-1874916e-20c8-11e8-8d4c-b80cbd2eff2b.jpg)

If not, I believe having a package _automatically_ install resources on installation is a violation of CRAN policies in most cases.  The [few exceptions](https://github.com/x13org/x13binary) I know of relate to license issues or platform-specific installs of compiled binaries.  But plenty of packages have functions to install system dependencies that users run _after_ install, such as [`keras::install_keras()`](https://keras.rstudio.com/reference/install_keras.html).  So having a `tika::install_tika()` function that downloads the `.jar` file (probably using [`rappdirs::user_data_dir()`](https://github.com/r-lib/rappdirs) to select the install path) is a reasonable alternative strategy.  In this case you will want to mark all examples as `\dontrun{}` and use `skip_on_cran()` for your tests so the package test properly on CRAN.  You can still run `install_tika()` and run the tests on Travis.
