### 2018-03-06 3701f2d
TIKA-2576 -- Upgrade commons compress and add detection and parsing of zstd (if user provides com.github.luben:zstd-jni... via Andreas Meier
