### 2018-03-07 #191 @noamross
I advise against two packages. It will be confusing for users make more effort for maintenance. I think an `install_tika()` function makes for a fairly simple experience - you can print a package startup message if the .jar if missing, and a similar informative error for any Tika-calling function.  Plus you can update the Tika version by one line of code - the download  URL - in the package. CRAN publication will maximize your user base, as this package will be a great useful to many people.

I'm OK with marking this as "accepted" now, but I would prefer that if the plan is to change the user experience in this way soon, we do so before we move forward with blog posts and such.
