### 2018-03-07 #191 @predict-r
Alright, I’ll make the install script. :)
