### 2018-03-07 b9e9e5b
TIKA-2594 -- improve eml detection for those starting with Subject: and containing html
