+++ tika-parsers/src/test/java/org/apache/tika/mime/TestMimeTypes.java
@@ -908,6 +908,9 @@ public class TestMimeTypes {
         //x- custom header
         assertTypeDetection("testRFC822_x-.eml", "message/rfc822");
 
+        //embedded xhtml and img
+        assertTypeDetection("testEML_embedded_xhtml_and_img.eml", "message/rfc822");
+
     }
     
     @Test
