### 2018-03-07 0903104
TIKA-2594 -- improve eml detection for those starting with Subject: and containing html
