### 2018-03-07 7e2b1e7
TIKA-2592 -- ignore charsets not supported by IANA in html meta-headers via Andreas Meier.
