### 2018-03-07 164c928
TIKA-2592 -- ignore charsets not supported by IANA in html meta-headers via Andreas Meier.
