+++ CHANGES.txt
@@ -5,6 +5,9 @@ Release 2.0.0 - ???
 
    Other changes
 
+   * Revert to listenForAllRecords=false in ExcelExtractor
+     via Grigoriy Alekseev (TIKA-2590)
+
    * Add workaround to identify TIFFs that might confuse
      commons-compress's tar detection via Daniel Schmidt
      (TIKA-2591)
