### 2018-03-07 9b7154c
TIKA-2527 -- Various new mimes and typo fixes in tika-mimetypes.xml via Andreas Meier.
