+++ CHANGES.txt
@@ -5,6 +5,9 @@ Release 2.0.0 - ???
 
    Other changes
 
+   * Various new mimes and typo fixes in tika-mimetypes.xml
+     via Andreas Meier (TIKA-2527).
+
    * Revert to listenForAllRecords=false in ExcelExtractor
      via Grigoriy Alekseev (TIKA-2590)
 
