### 2018-03-07 33f756f
TIKA-2527 -- Various new mimes and typo fixes in tika-mimetypes.xml via Andreas Meier.
