### 2018-03-08 c996d01
TIKA-2568: detection of full encrypted 7z files
