+++ tika-parsers/src/main/java/org/apache/tika/parser/pkg/PackageParser.java
@@ -252,10 +252,14 @@ public class PackageParser extends AbstractParser {
                 }
                 
                 SevenZFile sevenz;
-                if (password == null) {
-                    sevenz = new SevenZFile(tstream.getFile());
-                } else {
-                    sevenz = new SevenZFile(tstream.getFile(), password.getBytes("UnicodeLittleUnmarked"));
+                try{
+                    if (password == null) {
+                        sevenz = new SevenZFile(tstream.getFile());
+                    } else {
+                        sevenz = new SevenZFile(tstream.getFile(), password.getBytes("UnicodeLittleUnmarked"));
+                    }
+                }catch(PasswordRequiredException e){
+                    throw new EncryptedDocumentException(e);
                 }
                 
                 // Pending a fix for COMPRESS-269 / TIKA-1525, this bit is a little nasty
