### 2018-03-08 fcfd4e1
TIKA-2568: detection of full encrypted 7z files
