### 2018-03-08 a255df0
TIKA-2338 support for tif in pdfs
