### 2018-03-08 2e1a810
TIKA-2338 support for tif in pdfs
