+++ CHANGES.txt
@@ -1,5 +1,7 @@
 Release 1.18 - ???
    
+   * Support for TIFF images in PDF files (TIKA-2338)
+   
    * Detection of full encrypted 7z files (TIKA-2568)
 
    * Various new mimes and typo fixes in tika-mimetypes.xml
