### 2018-03-08 8ecfeef
TIKA-2530 -- temporary workaround -- check for zero length byte array in rtf body to avoid buffer underflow from POI, via Pascal Essiembre.
