### 2018-03-08 bbf8447
TIKA-2591 -- prevent AIOOBE when haystack shorter than needle
