### 2018-03-09 1df10c3
TIKA-2604 -- properly escape (or not) class path in windows and linux environments.
