### 2018-03-09 #191 @predict-r
I've implemented changes needed for `install_tika()` to work.

However, along the way, I ran into a big Apache Tika bug! It was in the batch processor. The good news is that I reported the bug, and it was immediately patched!  See [1].

A potential issue is the next major release of Tika with the patch, Tika version 1.18, will be released soon, but I'm not sure exactly when...I've asked.

In the meantime, what about having `install_tika()` get the patched development version directly from Apache? See [2].  This would come with warnings that it is a development version.

If people have thoughts on installing a development version of Tika, I'd like to hear them.

[1] https://issues.apache.org/jira/browse/TIKA-2604?page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel&focusedCommentId=16393363#comment-16393363
[2] https://builds.apache.org/job/tika-branch-1x/10/org.apache.tika$tika-app/

 P.S. I'm kinda proud of finding the bug!
