### 2018-03-10 #191 @noamross
👏 Good work finding that @predict-r! I had problems with another program crashing recently because of spaces in file paths, so I'm glad this bug is squashed.

We'll see what the response is from the developer, but [based on this email thread](https://lists.apache.org/list.html?dev@tika.apache.org:lte=1M:1.18) it seems that 1.18 is only a week or two off.  If it were _months_ away I would suggest an jumping ahead.  I would say it's fine to the patched version on GitHub now, (esp. given that `user_data_dir()` is likely to have spaces on many machines), but wait until 1.18 before CRAN submission.  GitHub packages are generally perceived to be "development" versions.  You can put a note about the patched version in the README.

I will take a careful look at the remaining changes to wrap this up later this weekend.
