### 2018-03-10 #191 @predict-r
Thanks ! The install_tika() is temporarily set to download the development version. The package is passing Travis.
