### 2018-03-12 #191 @noamross
Approved! Great work, and thanks a lot @juliasilge and @davidgohel for your reviews!  The `install_tika()` functions and docs are looking good, and I commend your good use of assertions and failsafes.

To-dos:
- [ ] Update badges, CI and coverage links, and URLs to point to the new repo address.

- If you would like to, and if your reviewers agree, you may choose to credit their reviews in the DESCRIPTION file by adding them as `'rev'`-type contributors ([example](https://github.com/ropensci/rrricanes/blob/master/DESCRIPTION#L11-L12))

Welcome aboard! I for one am quite excited about **rtika**. We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.
