### 2018-03-12 aa3134d
Begin implementing the Supplemental and Fallback parsers from https://wiki.apache.org/tika/CompositeParserDiscussion
