+++ tika-core/src/main/java/org/apache/tika/parser/multiple/AbstractMultipleParser.java
@@ -1,0 +1,231 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.tika.parser.multiple;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Path;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import org.apache.tika.exception.TikaException;
+import org.apache.tika.io.TemporaryResources;
+import org.apache.tika.io.TikaInputStream;
+import org.apache.tika.metadata.Metadata;
+import org.apache.tika.mime.MediaType;
+import org.apache.tika.mime.MediaTypeRegistry;
+import org.apache.tika.parser.AbstractParser;
+import org.apache.tika.parser.ParseContext;
+import org.apache.tika.parser.Parser;
+import org.apache.tika.parser.ParserDecorator;
+import org.xml.sax.ContentHandler;
+import org.xml.sax.SAXException;
+
+/**
+ * Abstract base class for parser wrappers which may / will
+ *  process a given stream multiple times, merging the results
+ *  of the various parsers used.
+ * End users should normally use {@link FallbackParser} or
+ *  {@link SupplementingParser} along with a Strategy.
+ *
+ * @since Apache Tika 1.18
+ */
+public abstract class AbstractMultipleParser extends AbstractParser {
+    /**
+     * Serial version UID.
+     */
+    private static final long serialVersionUID = 5383668090329836559L;
+
+    /**
+     * The various strategies for handling metadata emitted by
+     *  multiple parsers.
+     * Note that not all will be supported by all subclasses.
+     */
+    public enum MetadataPolicy {
+        /**
+         * Before moving onto another parser, throw away
+         *  all previously seen metadata
+         */
+        DISCARD_ALL,
+        /**
+         * The first parser to output a given key wins,
+         *  merge in non-clashing other keys
+         */
+        FIRST_WINS,
+        /**
+         * The last parser to output a given key wins,
+         *  overriding previous parser values for a
+         *  clashing key.
+         */
+        LAST_WINS,
+        /**
+         * Where multiple parsers output a given key,
+         *  store all their different (unique) values
+         */
+        KEEP_ALL
+    };
+    
+    // TODO Figure out some sort of Content Policy and how
+    //  it might possibly work
+
+    /**
+     * Media type registry.
+     */
+    private MediaTypeRegistry registry;
+    
+    /**
+     * How we should handle metadata clashes
+     */
+    private MetadataPolicy policy;
+    
+    /**
+     * List of the multiple parsers to try.
+     */
+    private List<Parser> parsers;
+    
+    /**
+     * Computed list of Mime Types to offer, which is all
+     *  those in common between the parsers.
+     * For explicit mimetypes only, use a {@link ParserDecorator}
+     */
+    private Set<MediaType> offeredTypes;
+    
+    // TODO Tika Config XML Support for these parsers and their
+    //  metadata policies + parsers + mimetypes
[…]
