### 2018-03-12 #191 @noamross
I made that change.  You should have repo settings control now.  All set!  I'm going to close this issue, but @stefaniebutland will respond here about blog timing.  Here are our editorial and technical guidelines for posts: https://github.com/ropensci/roweb2#contributing-a-blog-posts, and some other posts in the onboarding series https://ropensci.org/tags/review/
