### 2018-03-27 #191 @predict-r
I posted a draft here:

https://github.com/predict-r/roweb2/pull/1

It's a pull request on a fork...hope I did that correctly.
