### 2018-03-28 #191 @stefaniebutland
Thank you for doing this @predict-r.

> hope I did that correctly
It looks like you made a pull request to your own fork of our repo.

From https://github.com/predict-r/roweb2 you should create a pull request with branch `predict-r-patch-1` in your fork  to `master` on the ropensci repo. It will look like the image here

![screen shot 2018-03-27 at 3 09 51 pm](https://user-images.githubusercontent.com/11927811/38003585-01456ce4-31ed-11e8-819e-888e3370f082.png)

I have a few other blog posts to review first so I will get you some feedback likely early next week.
